Moved Message to the Model namespace
